7.5 클래스 추출하기 - 리팩터링 후
