+++ src/main/java/io/sjhello/refactoring/chapter7/extractclass/Person.java
@@ -3,17 +3,11 @@ package io.sjhello.refactoring.chapter7.extractclass;
 public class Person {
 
     private String name;
-    private String officeAreaCode;
-    private String officeNumber;
+    private TelephoneNumber telephoneNumber;
 
     public Person(String name, String officeAreaCode, String officeNumber) {
         this.name = name;
-        this.officeAreaCode = officeAreaCode;
-        this.officeNumber = officeNumber;
-    }
-
-    public String getTelephoneNumber() {
-        return officeAreaCode + "-" + officeNumber;
+        this.telephoneNumber = new TelephoneNumber(officeAreaCode, officeNumber);
     }
 
     public String getName() {
@@ -24,19 +18,7 @@ public class Person {
         this.name = name;
     }
 
-    public String getOfficeAreaCode() {
-        return officeAreaCode;
-    }
-
-    public void setOfficeAreaCode(String officeAreaCode) {
-        this.officeAreaCode = officeAreaCode;
-    }
-
-    public String getOfficeNumber() {
-        return officeNumber;
-    }
-
-    public void setOfficeNumber(String officeNumber) {
-        this.officeNumber = officeNumber;
+    public TelephoneNumber getTelephoneNumber() {
+        return telephoneNumber;
     }
 }
